requirements/base.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
sqlalchemy>=2.0.23
requests>=2.31.0
desktop-notifier>=3.4.3
backoff>=2.2.1
python-dotenv>=1.0.0
